move actionedState, expired and event_completed fields into status
